Next.js migration 1, move config out of state store
- Load config in server components and serialize by passing to client components.
- Move metadata loading down to body.
- App, Header can now be fully server rendered.
