+++ .env.local.example
@@ -5,3 +5,5 @@ REACT_APP_BC_BASE_MAP_TILES_URL=https://services.pacificclimate.org/tiles/bc-alb
 REACT_APP_YNWT_BASE_MAP_TILES_URL=https://services.pacificclimate.org/tiles/yukon-albers-lite/{z}/{x}/{y}.png
 
 PUBLIC_URL=http://localhost:3000/
+
+REACT_APP_APP_VERSION=local
