+++ src/app/page.js
@@ -1,4 +1,8 @@
 import App from "../components/main/App/App";
+
+import config from "../server-data/config";
+
 export default function Home() {
-  return <App />;
+  console.log(config);
+  return <App config={config} />;
 }
