+++ src/components/info/NetworksMetadata/__tests__/smoke.js
@@ -1,9 +1,10 @@
 import React from "react";
 import { createRoot } from "react-dom/client";
 import NetworksMetadata from "../NetworksMetadata";
+import config from "../../../../server-data/config";
 
 it("renders without crashing", () => {
   const container = document.createElement("div");
   const root = createRoot(container);
-  root.render(<NetworksMetadata allNetworks={[]} />);
+  root.render(<NetworksMetadata config={config} allNetworks={[]} />);
 });
